Worked on calendar API client functions.
