+++ Citizen-Science/src/api/auth.ts
@@ -1,4 +1,3 @@
-import { Event } from "../../../api/src/interfaces/Event";
 
 export const loginUser = async (email: string, password: string) => {
 	const response = await fetch(`${process.env.EXPO_PUBLIC_API_URL}/auth/login`, {
@@ -54,21 +53,3 @@ export const logoutUser = async (userId: number) => {
 		throw error;
 	}
 }
-
-export const createEvent = async (eventData: Event) => {
-	const response = await fetch(`${process.env.EXPO_PUBLIC_API_URL}/events/create`, {
-		method: 'POST',
-		headers: {
-			'Content-Type': 'application/json',
-		},
-		body: JSON.stringify(eventData),
-	});
-	if (!response.ok) {
-		const errorResponse = await response.json();
-		const error = new Error(errorResponse.error || response.statusText);
-		error.name = `HTTP Error ${response.status}`;
-		throw error;
-	}
-
-	return await response.json();
-};
